工单管理修订版

+++ ops-work-order-system/src/main/java/com/example/opsworkordersystem/controller/api/WorkOrderController.java
@@ -10,5 +10,5 @@ public interface WorkOrderController {
 
     List<WorkOrder> getWorkOrdersByStatus(String status);
 
-    List<WorkOrder> getWorkOrdersByCreator(Long createdById);
+    List<WorkOrder> getWorkOrdersByCreator(Integer createdById);
 }
